requirements.txt
aiogram
asyncpg
python-dotenv
apscheduler
aiohttp
beautifulsoup4
sqlalchemy
unidecode